11° commit: Cambios en css, en expositor, separacion de form-pedido, se agrego metodos en usuarioControlador, y creacion de PedidoControlador - 2/10/24

+++ FullStack-WebServicio/src/main/java/com/FinalEgg/ServiChacras/entidades/Cliente.java
@@ -25,5 +25,5 @@ public class Cliente {
     private List<Pedido> pedidos;
 
     @Column(name = "promedio_puntuacion")
-    private Integer promPuntuacion;
+    private Double promPuntuacion;
 }
